api/requirements-flask.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python-headless==4.8.1.78
insightface==0.7.3
numpy==1.24.3
Pillow==10.0.1
python-dateutil==2.8.2
onnxruntime==1.16.3
scikit-image==0.21.0